Fashion Project/requirements.txt
Flask
numpy
pandas
Pillow
scikit-learn
tensorflow==2.20.0
requests
gunicorn
